backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic[email]==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
cloudinary==1.41.0
resend==2.4.0
slowapi==0.1.9
python-magic==0.4.27
httpx==0.27.2
pytest==8.3.3
